+++ app/src/main/java/app/musikus/goals/presentation/GoalsUiEvent.kt
@@ -8,10 +8,14 @@
 
 package app.musikus.goals.presentation
 
-import app.musikus.goals.domain.GoalInstanceWithProgressAndDescriptionWithLibraryItems
 import app.musikus.core.domain.GoalsSortMode
+import app.musikus.goals.data.entities.GoalPeriodUnit
+import app.musikus.goals.data.entities.GoalType
+import app.musikus.goals.domain.GoalInstanceWithProgressAndDescriptionWithLibraryItems
+import app.musikus.library.data.daos.LibraryItem
 
 typealias GoalsUiEventHandler = (GoalsUiEvent) -> Unit
+typealias GoalDialogEventHandler = (GoalDialogUiEvent) -> Unit
 
 sealed class GoalsUiEvent {
     data object BackButtonPressed : GoalsUiEvent()
@@ -38,4 +42,15 @@ sealed class GoalsUiEvent {
     data object ClearActionMode : GoalsUiEvent()
 
     data class DialogUiEvent(val dialogEvent: GoalDialogUiEvent) : GoalsUiEvent()
+}
+
+sealed class GoalDialogUiEvent {
+    data class HourTargetChanged(val hours: Int?) : GoalDialogUiEvent()
+    data class MinuteTargetChanged(val minutes: Int?) : GoalDialogUiEvent()
+    data class PeriodChanged(val period: Int) : GoalDialogUiEvent()
+    data class PeriodUnitChanged(val periodUnit: GoalPeriodUnit) : GoalDialogUiEvent()
+    data class GoalTypeChanged(val goalType: GoalType) : GoalDialogUiEvent()
+    data class LibraryItemsSelected(val selectedLibraryItems: List<LibraryItem>) : GoalDialogUiEvent()
+    data object Confirm : GoalDialogUiEvent()
+    data object Dismiss : GoalDialogUiEvent()
 }
